requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pillow==9.5.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
sqlalchemy==2.0.20
alembic==1.11.1
psycopg2-binary==2.9.7
python-multipart==0.0.6
pytest==7.4.0
httpx==0.27.0
pytest-asyncio==0.22.0
